+++ docs/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 <!-- https://medium.com/@dtgasparyan/feature-sliced-design-the-ideal-frontend-architecture-84d701ad44ba -->
 ---
+### 21 APR 2025
+- Renamed `auth.validator` module to `login.validator`.
+- Moved `jwt.verify` from `auth.controller` to `authenticateToken` module.
+- Renamed `verifyToken` module to `authenticateToken`.
+- Added `updatedAt` field to `Post` model.
+---
 ### 20 APR 2025
 - Included `blog`, `posts` and `comments` for user objects returned from `getUser` and `getUsers` query functions.
 - Created `verifyToken` module and middleware.